test(chain): cover authority adapter failover
