+++ packages/chain/test/context-graph-authority-snapshot.unit.test.ts
@@ -6,12 +6,6 @@ import { ethers } from 'ethers';
 import type { ContextGraphAuthoritySnapshot } from '../src/chain-adapter.js';
 import { EVMChainAdapter } from '../src/evm-adapter.js';
 import { MockChainAdapter } from '../src/mock-adapter.js';
-import {
-  ContextGraphAuthorityHistoryCache,
-  resolveContextGraphAuthorityHistory,
-  type ContextGraphAuthorityHistoryCreationEvent,
-  type ResolveContextGraphAuthorityHistoryInput,
-} from '../src/context-graph-authority-history.js';
 
 const OWNER = '0x1111111111111111111111111111111111111111';
 const MEMBER = '0x2222222222222222222222222222222222222222';
@@ -25,7 +19,6 @@ const CREATION_HASH = `0x${'66'.repeat(32)}`;
 const POLICY_HASH = `0x${'77'.repeat(32)}`;
 const NEXT_POLICY_HASH = `0x${'78'.repeat(32)}`;
 const NAME_HASH = `0x${'88'.repeat(32)}`;
-const DIRECT_READ_SCOPE = {};
 
 function event(
   blockNumber: number,
@@ -34,50 +27,6 @@ function event(
   args: readonly unknown[] = [],
 ) {
   return { blockNumber, index, blockHash, args };
-}
-
-function directHistoryInput(params: Readonly<{
-  cache: ContextGraphAuthorityHistoryCache;
-  cacheKey: string;
-  blockNumber: number;
-  blockHash: string;
-  coldReads?: { count: number };
-  ordinaryRanges?: Array<readonly [number, number]>;
-  blockHashes?: Readonly<Record<number, string>>;
-  creationGate?: Readonly<{ entered(): void; wait: Promise<void> }>;
-  readScope?: object;
-  signal?: AbortSignal;
-  omitCreationNameHash?: boolean;
-}>): ResolveContextGraphAuthorityHistoryInput {
-  return {
-    cache: params.cache,
-    cacheKey: params.cacheKey,
-    readScope: params.readScope ?? DIRECT_READ_SCOPE,
-    contextGraphId: 9n,
-    finalized: { number: params.blockNumber, hash: params.blockHash },
-    pageSize: 100,
-    signal: params.signal,
-    loadColdFromBlock: async () => {
-      if (params.coldReads) params.coldReads.count += 1;
-      return 1;
-    },
-    readBlockHash: async (blockNumber) => params.blockHashes?.[blockNumber]
-      ?? (blockNumber === params.blockNumber ? params.blockHash : null),
-    readCreationEvents: async () => {
-      params.creationGate?.entered();
-      if (params.creationGate) await params.creationGate.wait;
-      return [{
-        blockNumber: 1,
-        blockHash: `0x${'01'.repeat(32)}`,
-        index: 0,
-        ...(params.omitCreationNameHash ? {} : { nameHash: NAME_HASH }),
-      }] as unknown as readonly ContextGraphAuthorityHistoryCreationEvent[];
-    },
-    readEvents: async (_query, fromBlock, toBlock) => {
-      params.ordinaryRanges?.push([fromBlock, toBlock]);
-      return [];
-    },
-  };
 }
 
 interface AuthorityEvidence {
@@ -338,6 +287,67 @@ describe('RFC-64 Context Graph authority snapshots', () => {
     await expect(pending).rejects.toThrow('anchor changed');
   });
 
+  it('lets adapter failover leave a stalled same-head history reader behind', async () => {
+    const { adapter: typedAdapter } = makeEvmAuthorityAdapter();
+    const adapter = typedAdapter as any;
+    const baseContract = adapter.contracts.contextGraphStorage.connect();
+    const historyEntered = Promise.withResolvers<void>();
+    const stalledHistory = new Promise<never>(() => {});
+    const makeProvider = () => ({
+      getBlock: async (tag: string | number) => ({
+        number: tag === 'finalized' ? 30 : Number(tag),
+        hash: FINALIZED_HASH,
+      }),
+      getNetwork: async () => ({ chainId: 31337n }),
+    });
+    const stalledProvider = makeProvider();
+    const healthyProvider = makeProvider();
+    const stalledContract = {
+      ...baseContract,
+      queryFilter: async () => {
+        historyEntered.resolve();
+        return stalledHistory;
+      },
+    };
+    const connectedProviders: object[] = [];
+    adapter.contracts.contextGraphStorage = {
+      connect: (provider: object) => {
+        connectedProviders.push(provider);
+        return provider === stalledProvider ? stalledContract : baseContract;
+      },
+    };
+    adapter.readTipProvider = async (
+      _label: string,
+      read: (provider: object) => Promise<unknown>,
+    ) => {
+      const stalledAttempt = read(stalledProvider);
+      void stalledAttempt.catch(() => {});
+      await historyEntered.promise;
+      return read(healthyProvider);
+    };
+
+    let timeout: ReturnType<typeof setTimeout> | undefined;
+    try {
+      const result = await Promise.race([
+        typedAdapter.getContextGraphAuthoritySnapshot(9n),
+        new Promise<never>((_resolve, reject) => {
+          timeout = setTimeout(
+            () => reject(new Error('healthy authority provider remained captured by stalled reader')),
+            1_000,
+          );
+        }),
+      ]);
+      expect(result).toMatchObject({
+        contextGraphId: '9',
+        nameHash: NAME_HASH,
+        sourceBlockNumber: '21',
+      });
+      expect(connectedProviders).toEqual([stalledProvider, healthyProvider]);
+    } finally {
+      if (timeout !== undefined) clearTimeout(timeout);
+    }
+  });
+
   it('discards the incremental watermark when its finalized anchor was replaced', async () => {
     const { adapter, evidence, advanceAuthorityHead, replaceCachedAnchor } =
       makeEvmAuthorityAdapter();
@@ -379,188 +389,6 @@ describe('RFC-64 Context Graph authority snapshots', () => {
       policyVersion: '4',
     });
     expect(harness.evidence.ranges.slice(18)).toEqual(Array(10).fill([31, 35]));
-  });
-
-  it('coalesces overlapping resolutions for the same finalized head', async () => {
-    const cache = new ContextGraphAuthorityHistoryCache();
-    const coldReads = { count: 0 };
-    const entered = Promise.withResolvers<void>();
-    const release = Promise.withResolvers<void>();
-    const input = directHistoryInput({
-      cache,
-      cacheKey: 'same-head',
-      blockNumber: 30,
-      blockHash: FINALIZED_HASH,
-      coldReads,
-      creationGate: { entered: entered.resolve, wait: release.promise },
-    });
-    const first = resolveContextGraphAuthorityHistory(input);
-    await entered.promise;
-    const second = resolveContextGraphAuthorityHistory(input);
-    release.resolve();
-    const [firstResolution, secondResolution] = await Promise.all([first, second]);
-    expect(coldReads.count).toBe(1);
-    expect(firstResolution.state).toBe(secondResolution.state);
-    await Promise.all([firstResolution.publish(), secondResolution.publish()]);
-  });
-
-  it('does not let a stalled provider attempt capture same-head failover', async () => {
-    const cache = new ContextGraphAuthorityHistoryCache();
-    const stalledScope = {};
-    const healthyScope = {};
-    const entered = Promise.withResolvers<void>();
-    const release = Promise.withResolvers<void>();
-    const stalled = resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache,
-      cacheKey: 'provider-failover',
-      blockNumber: 30,
-      blockHash: FINALIZED_HASH,
-      readScope: stalledScope,
-      creationGate: { entered: entered.resolve, wait: release.promise },
-    }));
-    await entered.promise;
-
-    const healthy = await resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache,
-      cacheKey: 'provider-failover',
-      blockNumber: 30,
-      blockHash: FINALIZED_HASH,
-      readScope: healthyScope,
-    }));
-    expect(healthy.state.nameHash).toBe(NAME_HASH);
-    await healthy.publish();
-
-    release.resolve();
-    await expect(stalled).resolves.toMatchObject({ state: healthy.state });
-  });
-
-  it('does not let one reader abort an independent same-head reader', async () => {
-    const cache = new ContextGraphAuthorityHistoryCache();
-    const readScope = {};
-    const firstAbort = new AbortController();
-    const secondAbort = new AbortController();
-    firstAbort.abort();
-
-    const cancelled = resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache,
-      cacheKey: 'independent-abort',
-      blockNumber: 30,
-      blockHash: FINALIZED_HASH,
-      readScope,
-      signal: firstAbort.signal,
-    }));
-    const healthy = resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache,
-      cacheKey: 'independent-abort',
-      blockNumber: 30,
-      blockHash: FINALIZED_HASH,
-      readScope,
-      signal: secondAbort.signal,
-    }));
-
-    await expect(cancelled).rejects.toMatchObject({ name: 'AbortError' });
-    await expect(healthy).resolves.toMatchObject({ state: { nameHash: NAME_HASH } });
-  });
-
-  it('retains a newer watermark when an older load publishes last', async () => {
-    const cache = new ContextGraphAuthorityHistoryCache();
-    const entered = Promise.withResolvers<void>();
-    const release = Promise.withResolvers<void>();
-    const older = resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache,
-      cacheKey: 'out-of-order',
-      blockNumber: 30,
-      blockHash: FINALIZED_HASH,
-      creationGate: { entered: entered.resolve, wait: release.promise },
-    }));
-    await entered.promise;
-    const newer = await resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache,
-      cacheKey: 'out-of-order',
-      blockNumber: 35,
-      blockHash: NEXT_FINALIZED_HASH,
-    }));
-    await newer.publish();
-    release.resolve();
-    await (await older).publish();
-
-    const ranges: Array<readonly [number, number]> = [];
-    const next = await resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache,
-      cacheKey: 'out-of-order',
-      blockNumber: 36,
-      blockHash: `0x${'57'.repeat(32)}`,
-      blockHashes: { 35: NEXT_FINALIZED_HASH },
-      ordinaryRanges: ranges,
-    }));
-    expect(ranges).toEqual(Array(5).fill([36, 36]));
-    await next.publish();
-  });
-
-  it('prevents clear from being undone by an in-flight load', async () => {
-    const cache = new ContextGraphAuthorityHistoryCache();
-    const coldReads = { count: 0 };
-    const entered = Promise.withResolvers<void>();
-    const release = Promise.withResolvers<void>();
-    const pending = resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache,
-      cacheKey: 'cleared',
-      blockNumber: 30,
-      blockHash: FINALIZED_HASH,
-      coldReads,
-      creationGate: { entered: entered.resolve, wait: release.promise },
-    }));
-    await entered.promise;
-    cache.clear();
-    release.resolve();
-    const stale = await pending;
-    await expect(stale.publish()).rejects.toThrow('invalidated');
-
-    const fresh = await resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache,
-      cacheKey: 'cleared',
-      blockNumber: 30,
-      blockHash: FINALIZED_HASH,
-      coldReads,
-    }));
-    await fresh.publish();
-    expect(coldReads.count).toBe(2);
-  });
-
-  it('bounds authority history with LRU retention', async () => {
-    const cache = new ContextGraphAuthorityHistoryCache(2);
-    const coldReads = new Map<string, { count: number }>();
-    const publish = async (cacheKey: string) => {
-      const counter = coldReads.get(cacheKey) ?? { count: 0 };
-      coldReads.set(cacheKey, counter);
-      const resolution = await resolveContextGraphAuthorityHistory(directHistoryInput({
-        cache,
-        cacheKey,
-        blockNumber: 30,
-        blockHash: FINALIZED_HASH,
-        coldReads: counter,
-      }));
-      await resolution.publish();
-    };
-    await publish('a');
-    await publish('b');
-    await publish('a');
-    await publish('c');
-    expect(cache.size).toBe(2);
-    await publish('b');
-    expect(coldReads.get('a')?.count).toBe(1);
-    expect(coldReads.get('b')?.count).toBe(2);
-    expect(coldReads.get('c')?.count).toBe(1);
-  });
-
-  it('rejects a malformed creation event without a name hash at runtime', async () => {
-    await expect(resolveContextGraphAuthorityHistory(directHistoryInput({
-      cache: new ContextGraphAuthorityHistoryCache(),
-      cacheKey: 'malformed-creation',
-      blockNumber: 30,
-      blockHash: FINALIZED_HASH,
-      omitCreationNameHash: true,
-    }))).rejects.toThrow('creation event has no name hash');
   });
 
   it('provides the same authority surface in offline mock-chain mode', async () => {
